solving warnings for building with Xcode 9

+++ Library/Categories/CoreData/Import/MagicalImportFunctions.m
@@ -67,9 +67,9 @@ UIColor *MRColorFromString(NSString *serializedColor)
         return nil;
     }
 
-    UIColor *color = [UIColor colorWithRed:(componentValues[ 0 ] / 255.0f)
-                                     green:(componentValues[ 1 ] / 255.0f)
-                                      blue:(componentValues[ 2 ] / 255.0f)
+    UIColor *color = [UIColor colorWithRed:(componentValues[ 0 ] / (CGFloat)255.)
+                                     green:(componentValues[ 1 ] / (CGFloat)255.)
+                                      blue:(componentValues[ 2 ] / (CGFloat)255.)
                                      alpha:componentValues[ 3 ]];
 
     free(componentValues);
@@ -86,9 +86,9 @@ NSColor *MRColorFromString(NSString *serializedColor)
         return nil;
     }
 
-    NSColor *color = [NSColor colorWithDeviceRed:(componentValues[ 0 ] / 255.0f)
-                                           green:(componentValues[ 1 ] / 255.0f)
-                                            blue:(componentValues[ 2 ] / 255.0f)
+    NSColor *color = [NSColor colorWithDeviceRed:(componentValues[ 0 ] / (CGFloat)255.)
+                                           green:(componentValues[ 1 ] / (CGFloat)255.)
+                                            blue:(componentValues[ 2 ] / (CGFloat)255.)
                                            alpha:componentValues[ 3 ]];
     free(componentValues);
     return color;
